+++ test/clj/rems/poller/test_email.clj
@@ -67,6 +67,11 @@
                               :field/optional false
                               :field/type :description}]}))
 
+(def ^:private get-license
+  (constantly {:id 1234
+               :licensetype "text"
+               :textcontent "foobar"}))
+
 (defn ^:private get-nothing [& _]
   nil)
 
@@ -81,7 +86,7 @@
                         (model/enrich-with-injections {:get-workflow get-workflow
                                                        :get-catalogue-item get-catalogue-item
                                                        :get-form-template get-form-template
-                                                       :get-license get-nothing
+                                                       :get-license get-license
                                                        :get-user get-nothing
                                                        :get-users-with-role get-nothing
                                                        :get-attachments-for-application get-nothing}))]
@@ -106,7 +111,8 @@
                       :application/field-values {1 "Application title"}}
                      {:application/id 7
                       :event/type :application.event/submitted
-                      :event/actor "applicant"}]]
+                      :event/actor "applicant"
+                      :event/time 13}]]
     (let [events (into
                   base-events
                   [{:application/id 7
@@ -126,11 +132,19 @@
                     :event/type :application.event/commented
                     :event/actor "commenter2"
                     :application/request-id "r1"
-                    :application/comment ["this is a comment"]}
+                    :application/comment "this is a comment"}
+                   {:application/id 7
+                    :event/type :application.event/remarked
+                    :event/actor "remarker"
+                    :application/comment "remark!"}
                    {:application/id 7
                     :event/type :application.event/member-added
                     :event/actor "handler"
                     :application/member {:userid "member"}}
+                   {:application/id 7
+                    :event/type :application.event/licenses-added
+                    :event/actor "handler"
+                    :application/licenses [{:license/id 1234}]}
                    {:application/id 7
                     :event/type :application.event/decision-requested
                     :event/actor "assistant"
@@ -148,64 +162,79 @@
                     :event/actor "assistant"}])]
       (is (= [[]
               []
-              [{:to-user "assistant",
-                :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                :body "Dear assistant,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "handler",
-                :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
+              [{:to-user "assistant"
+                :subject "A new application has been submitted (2001/3, \"Application title\")"
+                :body "Dear assistant,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "handler"
+                :subject "A new application has been submitted (2001/3, \"Application title\")"
+                :body "Dear Hannah Handler,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
               [{:to "somebody@example.com",
                 :subject "Invitation to participate in an application",
-                :body "Hello,\n\nThis email address (somebody@example.com) has been invited to participate in an application.\n\nParticipate: http://example.com/accept-invitation?token=abc\n\nPlease don't reply to this automatically generated message."}]
-              [{:to-user "commenter1",
-                :subject "Review requested (Hannah Handler: 2001/3, \"Application title\")",
-                :body "Dear commenter1,\n\nHannah Handler has requested a review on application 2001/3, \"Application title\".\n\nComment: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "commenter2",
-                :subject "Review requested (Hannah Handler: 2001/3, \"Application title\")",
-                :body "Dear commenter2,\n\nHannah Handler has requested a review on application 2001/3, \"Application title\".\n\nComment: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
+                :body "Dear Some Body,\n\nYou have been invited to participate in an application submitted by Alice Applicant. The title of the application is 2001/3, \"Application title\".\n\nYou can view the application and accept the terms of use: http://example.com/accept-invitation?token=abc\n\nPlease do not reply to this automatically generated message."}]
+              [{:to-user "commenter1"
+                :subject "Review request (2001/3, \"Application title\")"
+                :body "Dear commenter1,\n\nHannah Handler has requested your review on application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can review the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "commenter2"
+                :subject "Review request (2001/3, \"Application title\")"
+                :body "Dear commenter2,\n\nHannah Handler has requested your review on application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can review the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
               []
-              [{:to-user "assistant",
-                :subject "Review added (commenter2: 2001/3, \"Application title\")",
-                :body "Dear assistant,\n\ncommenter2 has reviewed application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "handler",
-                :subject "Review added (commenter2: 2001/3, \"Application title\")",
-                :body "Dear Hannah Handler,\n\ncommenter2 has reviewed application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
+              [{:to-user "assistant"
+                :subject "Application has been reviewed (2001/3, \"Application title\")"
+                :body "Dear assistant,\n\ncommenter2 has reviewed the application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the review: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "handler"
+                :subject "Application has been reviewed (2001/3, \"Application title\")"
+                :body "Dear Hannah Handler,\n\ncommenter2 has reviewed the application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the review: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+              [{:to-user "assistant"
+                :subject "Application has been commented on (2001/3, \"Application title\")"
+                :body "Dear assistant,\n\nremarker has commented on the application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the comment: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "handler"
+                :subject "Application has been commented on (2001/3, \"Application title\")"
+                :body "Dear Hannah Handler,\n\nremarker has commented on the application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the comment: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
               [{:to-user "member",
                 :subject "Added as a member of an application (2001/3, \"Application title\")",
-                :body "Dear member,\n\nYou've been added as a member of application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
-              [{:to-user "decider",
-                :subject "Decision requested (assistant: 2001/3, \"Application title\")",
-                :body "Dear decider,\n\nassistant has requested your decision on application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
-              [{:to-user "assistant",
-                :subject "Decision made (decider: 2001/3, \"Application title\")",
-                :body "Dear assistant,\n\ndecider has made a decision on application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "handler",
-                :subject "Decision made (decider: 2001/3, \"Application title\")",
-                :body "Dear Hannah Handler,\n\ndecider has made a decision on application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
+                :body "Dear member,\n\nYou've been added as a member of application 2001/3, \"Application title\".\n\nView application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
               [{:to-user "applicant"
-                :subject "Application approved (2001/3, \"Application title\")",
-                :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
+                :subject "New terms of use waiting for approval (2001/3, \"Application title\")"
+                :body "Dear Alice Applicant,\n\nHannah Handler has requested your acceptance for new terms of use for application 2001/3, \"Application title\".\n\nYou can view the application and accept the terms of use: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "member"
+                :subject "New terms of use waiting for approval (2001/3, \"Application title\")"
+                :body "Dear member,\n\nHannah Handler has requested your acceptance for new terms of use for application 2001/3, \"Application title\".\n\nYou can view the application and accept the terms of use: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "somebody"
+                :subject "New terms of use waiting for approval (2001/3, \"Application title\")"
+                :body "Dear somebody,\n\nHannah Handler has requested your acceptance for new terms of use for application 2001/3, \"Application title\".\n\nYou can view the application and accept the terms of use: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+              [{:to-user "decider",
+                :subject "Decision requested (2001/3, \"Application title\")",
+                :body "Dear decider,\n\nassistant has requested your decision on application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+              [{:to-user "assistant",
+                :subject "Decision notification (2001/3, \"Application title\")",
+                :body "Dear assistant,\n\ndecider has sent a decision on application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "handler",
+                :subject "Decision notification (2001/3, \"Application title\")",
+                :body "Dear Hannah Handler,\n\ndecider has sent a decision on application 2001/3, \"Application title\" submitted by Alice Applicant.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+              [{:to-user "applicant"
+                :subject "Your application has been approved (2001/3, \"Application title\")"
+                :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
                {:to-user "assistant"
                 :subject "Application approved (2001/3, \"Application title\")"
-                :body "Dear assistant,\n\nHannah Handler has approved the application 2001/3, \"Application title\" from Alice Applicant.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "member",
-                :subject "Application approved (2001/3, \"Application title\")",
-                :body "Dear member,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "somebody",
-                :subject "Application approved (2001/3, \"Application title\")",
-                :body "Dear somebody,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
+                :body "Dear assistant,\n\nHannah Handler has approved the application 2001/3, \"Application title\" from Alice Applicant.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "member"
+                :subject "Your application has been approved (2001/3, \"Application title\")"
+                :body "Dear member,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "somebody"
+                :subject "Your application has been approved (2001/3, \"Application title\")"
+                :body "Dear somebody,\n\nYour application 2001/3, \"Application title\" has been approved.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
               [{:to-user "applicant"
-                :subject "Application closed (2001/3, \"Application title\")",
-                :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
+                :subject "Your application has been closed (2001/3, \"Application title\")"
+                :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
                {:to-user "handler"
                 :subject "Application closed (2001/3, \"Application title\")"
-                :body "Dear Hannah Handler,\n\nassistant has closed the application 2001/3, \"Application title\" from Alice Applicant.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "member",
-                :subject "Application closed (2001/3, \"Application title\")",
-                :body "Dear member,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "somebody",
-                :subject "Application closed (2001/3, \"Application title\")",
-                :body "Dear somebody,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]]
+                :body "Dear Hannah Handler,\n\nassistant has closed the application 2001/3, \"Application title\" from Alice Applicant.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "member"
+                :subject "Your application has been closed (2001/3, \"Application title\")"
+                :body "Dear member,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "somebody"
+                :subject "Your application has been closed (2001/3, \"Application title\")"
+                :body "Dear somebody,\n\nYour application 2001/3, \"Application title\" has been closed.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]]
              (events-to-emails events))))
     (testing "application rejected"
       (let [events (conj base-events
@@ -214,38 +243,38 @@
                           :event/actor "handler"})]
         (is (= [[]
                 []
-                [{:to-user "assistant",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear assistant,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-                 {:to-user "handler",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
-                [{:subject "Application rejected (2001/3, \"Application title\")",
-                  :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been rejected.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message.",
-                  :to-user "applicant"}
-                 {:subject "Application rejected (2001/3, \"Application title\")"
-                  :body "Dear assistant,\n\nHannah Handler has rejected the application 2001/3, \"Application title\" from Alice Applicant.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."
-                  :to-user "assistant"}]]
+                [{:to-user "assistant"
+                  :subject "A new application has been submitted (2001/3, \"Application title\")"
+                  :body "Dear assistant,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+                 {:to-user "handler"
+                  :subject "A new application has been submitted (2001/3, \"Application title\")"
+                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+                [{:to-user "applicant"
+                  :subject "Your application has been rejected (2001/3, \"Application title\")",
+                  :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been rejected.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+                 {:to-user "assistant"
+                  :subject "Application rejected (2001/3, \"Application title\")",
+                  :body "Dear assistant,\n\nHannah Handler has rejected the application 2001/3, \"Application title\" from Alice Applicant.\n\nYou can view the application and the decision: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]]
                (events-to-emails events)))))
     (testing "id field can be overrided"
       (with-redefs [rems.config/env (assoc rems.config/env :application-id-column :id)]
         (is (= [[]
                 []
                 [{:to-user "assistant"
-                  :subject "Application submitted (Alice Applicant: 7, \"Application title\")"
-                  :body "Dear assistant,\n\nAlice Applicant has submitted an application (7, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
+                  :subject "A new application has been submitted (7, \"Application title\")"
+                  :body "Dear assistant,\n\nAlice Applicant has submitted a new application 7, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
                  {:to-user "handler"
-                  :subject "Application submitted (Alice Applicant: 7, \"Application title\")"
-                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (7, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]]
+                  :subject "A new application has been submitted (7, \"Application title\")"
+                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted a new application 7, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]]
                (events-to-emails base-events)))))
     (testing "application title is optional"
       (is (= [[]
-              [{:to-user "assistant",
-                :subject "Application submitted (Alice Applicant: 2001/3)",
-                :body "Dear assistant,\n\nAlice Applicant has submitted an application (2001/3): en title 11.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-               {:to-user "handler",
-                :subject "Application submitted (Alice Applicant: 2001/3)",
-                :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (2001/3): en title 11.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]]
+              [{:to-user "assistant"
+                :subject "A new application has been submitted (2001/3)"
+                :body "Dear assistant,\n\nAlice Applicant has submitted a new application 2001/3 to access resource(s) en title 11.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+               {:to-user "handler"
+                :subject "A new application has been submitted (2001/3)"
+                :body "Dear Hannah Handler,\n\nAlice Applicant has submitted a new application 2001/3 to access resource(s) en title 11.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]]
              (events-to-emails [{:application/id 7
                                  :application/external-id "2001/3"
                                  :event/type :application.event/created
@@ -257,7 +286,7 @@
                                 {:application/id 7
                                  :event/type :application.event/submitted
                                  :event/actor "applicant"}]))))
-    (testing "returning application to applicant and re-submitting"
+    (testing "returning application to applicant and resubmitting"
       (let [events (conj base-events
                          {:application/id 7
                           :event/type :application.event/returned
@@ -268,22 +297,22 @@
                           :event/actor "applicant"})]
         (is (= [[]
                 []
-                [{:to-user "assistant",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear assistant,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-                 {:to-user "handler",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
-                [{:to-user "applicant",
-                  :subject "Application returned (2001/3, \"Application title\")",
-                  :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been returned.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-                 {:to-user "assistant",
-                  :subject "Application returned (2001/3, \"Application title\")",
-                  :body "Dear assistant,\n\nHannah Handler has returned the application 2001/3, \"Application title\" from Alice Applicant.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]
-                [{:to-user "assistant",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear assistant,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}
-                 {:to-user "handler",
-                  :subject "Application submitted (Alice Applicant: 2001/3, \"Application title\")",
-                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted an application (2001/3, \"Application title\"): en title 11, en title 21.\n\nView application: http://example.com/#/application/7\n\nPlease don't reply to this automatically generated message."}]]
+                [{:to-user "assistant"
+                  :subject "A new application has been submitted (2001/3, \"Application title\")"
+                  :body "Dear assistant,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+                 {:to-user "handler"
+                  :subject "A new application has been submitted (2001/3, \"Application title\")"
+                  :body "Dear Hannah Handler,\n\nAlice Applicant has submitted a new application 2001/3, \"Application title\" to access resource(s) en title 11, en title 21.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+                [{:to-user "applicant"
+                  :subject "Your application has been returned for modifications (2001/3, \"Application title\")"
+                  :body "Dear Alice Applicant,\n\nYour application 2001/3, \"Application title\" has been returned to you for modifications.\n\nYou can modify and resubmit the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+                 {:to-user "assistant"
+                  :subject "Application has been returned for modifications (2001/3, \"Application title\")"
+                  :body "Dear assistant,\n\nHannah Handler has returned the application 2001/3, \"Application title\" to the applicant Alice Applicant for modifications.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]
+                [{:to-user "assistant"
+                  :subject "Application has been resubmitted (2001/3, \"Application title\")"
+                  :body "Dear assistant,\n\nApplication 2001/3, \"Application title\" has been resubmitted by Alice Applicant.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}
+                 {:to-user "handler"
+                  :subject "Application has been resubmitted (2001/3, \"Application title\")"
+                  :body "Dear Hannah Handler,\n\nApplication 2001/3, \"Application title\" has been resubmitted by Alice Applicant.\n\nYou can view the application: http://example.com/#/application/7\n\nPlease do not reply to this automatically generated message."}]]
                (events-to-emails events)))))))